small update

+++ src/store/voiceChat/actions.ts
@@ -1,32 +1,38 @@
 import { RootState } from "@/types/store";
 import { ActionTree } from "vuex";
-import { Group, TextChat, Message, Participant } from "@/types";
+import { Participant } from "@/types";
 import { voiceChatActionTypes } from "@/types/store";
-import { addParticipant, deleteParticipant, getVoiceChat, setMuted } from "@/FB_Queries/voiceChat";
+import { addParticipant, deleteParticipant, getVoiceChat, setMuted } from "@/queries/voiceChat";
 
 const voiceChatActions: ActionTree<RootState, RootState> = {
-  [voiceChatActionTypes.SET_VOICE_CHAT]: async ({ rootState }, { groupId, chatId, audioTrack }) => {
+  [voiceChatActionTypes.SET_VOICE_CHAT]: async ({ rootState }, { groupId, chatId }) => {
     await getVoiceChat(groupId, chatId, rootState);
   },
   [voiceChatActionTypes.ADD_PARTICIPANT]: async ({ rootState }, { groupId, chatId, UID }) => {
-    const newParticipant: Participant = {
-      userInfo: {
-        displayName: rootState.user.displayName || "",
-        photoURL: rootState.user.photoURL || "",
-        uid: rootState.user.uid,
-      },
-      joinedAt: Date.now(),
-      uid: String(UID),
-      muted: false,
-    };
+    if (rootState.user) {
+      const newParticipant: Participant = {
+        userInfo: {
+          displayName: rootState.user.displayName || "",
+          photoURL: rootState.user.photoURL || "",
+          uid: rootState.user.uid,
+        },
+        joinedAt: Date.now(),
+        uid: String(UID),
+        muted: false,
+      };
 
-    await addParticipant(groupId, chatId, newParticipant);
+      await addParticipant(groupId, chatId, newParticipant);
+    }
   },
   [voiceChatActionTypes.PARTICIPANT_LEFT]: ({ rootState }, { groupId, chatId }) => {
-    deleteParticipant(groupId, chatId, rootState.user.uid);
+    if (rootState.user) {
+      deleteParticipant(groupId, chatId, rootState.user.uid);
+    }
   },
   [voiceChatActionTypes.SET_MUTED]: ({ rootState }, { groupId, chatId, muted }) => {
-    setMuted(groupId, chatId, rootState.user.uid, muted);
+    if (rootState.user) {
+      setMuted(groupId, chatId, rootState.user.uid, muted);
+    }
   },
 };
